code review: move ApplySpanSampling into sampler.go
